requirements.txt
pyTelegramBotAPI==4.6.0
aiohttp==3.8.1
psycopg2-binary==2.9.3